app/requirements.txt
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
certifi==2018.4.16
chardet==3.0.4
decorator==4.3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.4
django-cors-headers==3.1.0
django-environ==0.4.5
django-filter==2.2.0
django-heroku==0.3.1
django-phone-field==1.8.0
djangorestframework==3.10.2
djangorestframework-simplejwt==4.3.0
entrypoints==0.3
fastparquet==0.3.2
gunicorn==19.9.0
idna==2.7
ipykernel==5.1.2
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.12.1
Jinja2==2.10.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==5.3.1
jupyter-console==6.0.0
jupyter-core==4.5.0
llvmlite==0.29.0
Markdown==3.1.1
MarkupSafe==1.1.1
mistune==0.8.4
mysqlclient==1.3.13
nbconvert==5.6.0
nbformat==4.4.0
notebook==6.0.0
numba==0.45.1
numpy==1.17.0
pandas==0.25.0
pandocfilters==1.4.2
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.4
prometheus-client==0.7.1
prompt-toolkit==2.0.9
psycopg2==2.8.3
ptyprocess==0.6.0
Pygments==2.2.0
PyJWT==1.7.1
pyrsistent==0.15.4
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
pyzmq==18.1.0
qtconsole==4.5.2
requests==2.19.1
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
thrift==0.11.0
tornado==6.0.3
traitlets==4.3.2
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
whitenoise==4.1.3
widgetsnbextension==3.5.1
